Split up release and release-candidate stages (#107)
* Change `run-local-registry` to run when needed

* Split up release and release-candidate stages

As we need to upload local artifacts on merge to master, but only
require promote existing images on tag

+++ Makefile
@@ -32,7 +32,10 @@ release: release-all
 check-style: check-style-all
 
 run-local-registry:
+	@echo "== run-local-registry"
+ifeq (, $(shell  docker ps --filter=name="dind-registry" --format="{{.Names}}"))
 	docker run -d --name=dind-registry --rm -p 5000:5000 registry:2
+endif
 
 check-system-dependencies:
 	@echo "== check-system-dependencies"
